+++ Sources/Navigator/PDF/PDFNavigatorViewController.swift
@@ -455,7 +455,7 @@ open class PDFNavigatorViewController:
         }
 
         if currentResourceIndex != index {
-            guard let document = PDFDocument(url: url.url) else {
+            guard let document = await makeDocument(at: url) else {
                 log(.error, "Can't open PDF document at \(url)")
                 return false
             }
@@ -481,6 +481,13 @@ open class PDFNavigatorViewController:
         }
 
         return true
+    }
+
+    private func makeDocument(at url: AbsoluteURL) async -> PDFKit.PDFDocument? {
+        let task = Task.detached(priority: .userInitiated) {
+            PDFDocument(url: url.url)
+        }
+        return await task.value
     }
 
     /// Updates the scale factors to match the currently visible pages.